requirements.txt
fastapi
uvicorn[standard]
youtube-transcript-api
llama-index
faiss-cpu
fastembed
google-cloud-storage
httpx
aiofiles
python-dotenv
